SignalScheduler: Support `configuration::Thread::All` mode
